Improve handling of Platform code
Improves the handling of platform specific code
by making better use of C++11.
